refactoring; mainly moving tests and methods to new directories and files
